+++ src/levels/Level.cpp
@@ -6,8 +6,8 @@
 Level::Level() {  }
 
 uint32_t Level::getStartPos()                               { return _startPos; }
-uint32_t Level::getWidth()                                  { return _width; }
-uint32_t Level::getHeight()                                 { return _height; }
+uint8_t Level::getWidth()                                   { return _width; }
+uint8_t Level::getHeight()                                  { return _height; }
 const Item * Level::getDoors()                              { return _doors; }
 
 const uint8_t * Level::getLevel()                           { return _level; }
@@ -18,13 +18,13 @@ char * Level::getTitleLine2()                               { return _titleLine2
 
 void Level::setDoors(const Item *value)                     { _doors = value; }
 void Level::setStartPos(const uint32_t value)               { _startPos = value; }
-void Level::setWidth(const uint32_t value)                  { _width = value; }
-void Level::setHeight(const uint32_t value)                 { _height = value; }
+void Level::setWidth(const uint8_t value)                   { _width = value; }
+void Level::setHeight(const uint8_t value)                  { _height = value; }
 
 void Level::setLevel(const uint8_t *value)                  { _level = value; }
 void Level::setMapTiles(const uint8_t * const *value)       { _map_tiles = value; }
   
-MapElement Level::getMapElement(uint32_t x, uint32_t y) {
+MapElement Level::getMapElement(uint8_t x, uint8_t y) {
 
 
   // Is there a door in this location?
@@ -51,7 +51,7 @@ MapElement Level::getMapElement(uint32_t x, uint32_t y) {
   uint8_t tileNumber = pgm_read_byte(&_level[_startPos + (x / MAP_TILE_WIDTH) + ((y / MAP_TILE_HEIGHT) * _width)]);
 
   const uint8_t *tile = _map_tiles[tileNumber];
-  uint16_t mapElement = pgm_read_byte(&tile[(x % MAP_TILE_WIDTH) + (((y % MAP_TILE_HEIGHT) / 8) * MAP_TILE_PHYSICAL_WIDTH)]) & (1 << (y % MAP_TILE_HEIGHT % 8));
+  uint8_t mapElement = pgm_read_byte(&tile[(x % MAP_TILE_WIDTH) + (((y % MAP_TILE_HEIGHT) / 8) * MAP_TILE_PHYSICAL_WIDTH)]) & (1 << (y % MAP_TILE_HEIGHT % 8));
   
   return (mapElement > 0 ? MapElement::Wall : MapElement::Floor);
